Scintilla 3.7.0 integration

+++ scintilla/lexers/LexErrorList.cxx
@@ -320,12 +320,12 @@ static void ColouriseErrorListLine(
 		int portionStyle = style;
 		while (const char *startSeq = strstr(linePortion, CSI)) {
 			if (startSeq > linePortion) {
-				styler.ColourTo(startPortion + (startSeq - linePortion), portionStyle);
+				styler.ColourTo(startPortion + static_cast<int>(startSeq - linePortion), portionStyle);
 			}
 			const char *endSeq = startSeq + 2;
 			while (!SequenceEnd(*endSeq))
 				endSeq++;
-			const int endSeqPosition = startPortion + (endSeq - linePortion) + 1;
+			const int endSeqPosition = startPortion + static_cast<int>(endSeq - linePortion) + 1;
 			switch (*endSeq) {
 			case 0:
 				styler.ColourTo(endPos, SCE_ERR_ESCSEQ_UNKNOWN);
